+++ src/app/app.component.ts
@@ -53,6 +53,7 @@ export class AppComponent implements OnInit{
     this.setTranslations();
     this.loadCharacters();
     this.loadGiftData();
+    this.loadPowersData();
   }
 
   private setTranslations() {
@@ -77,6 +78,13 @@ export class AppComponent implements OnInit{
     this.jsonService.getGiftsDataFromJson(FILES.GIFTS)
     .subscribe(data => {
       this.characterService.setGiftData(data);
+    });
+  }
+
+  private loadPowersData(){
+    this.jsonService.getPowersDataFromJson(FILES.POWERS)
+    .subscribe(data => {
+      this.characterService.setPowersData(data);
     });
   }
 }