+++ src/app/constants/number.constants.ts
@@ -4,6 +4,7 @@ export const MAGIC_NUMBERS = {
     N_1: 1,
     N_2: 2,
     N_3: 3,
+    N_4: 4,
     N_5: 5,
     N_6: 6,
     N_9: 9,