+++ src/app/services/json.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { map } from 'rxjs/operators';
 import { GiftData } from '../model/giftData';
+import { PowersData } from '../model/powerData';
 
 @Injectable({
   providedIn: 'root'
@@ -14,6 +15,14 @@ export class JsonService {
     return this.http.get(jsonFile).pipe(
       map(data => {
         return data as GiftData;
+      })
+    );
+  }
+
+  public getPowersDataFromJson(jsonFile: string) {
+    return this.http.get(jsonFile).pipe(
+      map(data => {
+        return data as PowersData;
       })
     );
   }